refactor: moved mqtt stuff to mqtt.js file
